Possibility to ignore nans when smoothing matrices
It is now possible to ignore nans when smoothing a matrix, if the nan value are present in the full  row/column. If there are isolated nan values they will be left as is and will thus influence the smoothing process.
The modifications to the two other functions is just to parametrize this new option.

TODO: if a big chunk of nans (many rows/cols) are removed, we could add padding to reduce border effects.